Removed messages_inbound and inbound property.
Messages_inbound was created under the assumption that a message frame
consists of 3 frames, as this may be untrue, the value provided by this
function is questionable at best.

I also removed the property for inbound, but the internal queue can
still be accessed using the private variable _inbound if needed, but I
deemed that unlikely.